introduce WorkingTimeService interface
